Introducing abstract class for the decompilers
